plugin system is now ASDF based
plugins are now defined by a file, plugin.asd, which tells wookie about
the plugin. this file contains a `defplugin` definition, which is
EXACTLY like asdf:defsystem (in fact, it wraps it). this macro is needed
to do some plugin name <--> asdf system matching, so use it or else.

plugin dependencies are now stored and resolved via ASDF. dependencies
are also read from ASDF when doing `unload-plugin` (now exported). this
reduces a lot of work on wookies part as far as plugin management goes,
since a lot of the work earlier was just duplicating ASDF. now there is
a good intersection, i think.

there is a minor change to logging: wookie-config now exports
*log-output* which is meant to hold a stream. if nil, output goes to
STDOuT.