add smart close to send-response/finish-response
unless specified, the request headers are now read in send-response and
finish-response to determine whether or not the connection should close.
this makes it so the app doesn't need to worry about it, and presistent
(or one-off) connections are handled/abstracted completely by wookie.

also, finish-response now takes the response object *instead of* the
chunked-stream object.