Adding :buffer-body option to defroute
Allows a route to specify that if the chunks start coming in before the
with-chunking handle is called, the body chunks be buffered into memory
until with-chunking has a chance to be called.

This can be useful if :pre-route returns a future and chunking starts
happening before it finishes, and the client doesn't wait around for
your 100 Continue.

It's also useful when you set up a chunked route but the client sends
all data in one dump (no chunking) AND your :pre-route delays on a
future. This way, your route's with-chunking handler will be called
(with the full body) directly after it is set up.

working great.