Better error handling in main parser
Fixes issue #34. When an error occurs while parsing headers, it is sent
off to the main-error-handler function, and the request is marked with
an error so the finished/body callbacks no longer process it.

When this happens, it's important for the app itself to send back an
error notice to the client, as Wookie will not do this for you.