Don't ignore explicitly-passed NIL response bodies.
It's a common impulse to pass NIL for responses that have no body (e.g.
404 response, or a session heartbeat url), but wookie currently ignores
them and doesn't set content-length. Without content-length and in the
absence of other headers, the browser will wait for a connection close to
end the response. If the server doesn't close the connection, the browser
hangs.

Instead, if the body has been explicitly supplied, always set a
content-length. This allows NIL to be used for empty bodies, without
breaking START-RESPONSE's use of SEND-RESPONSE to set headers.