Inherit from ERROR in WOOKIE-ERROR.
Any condition that represents a situation that must be handled callers
should inherit from at least SERIOUS-CONDITION; in this case, since the
condition represents an actual error, ERROR is appropriate. Without this,
code that handles ERROR or SERIOUS-CONDITION will miss any condition that
inherits from this -- for example, signalling WOOKIE-ERROR from a clack
handler bypasses clack's error-catching mechanism and invokes the debugger
in the server.