Defer all use of quicklisp symbols to run-time.
Previously, if quicklisp was available at compile time, these macros would
emit symbols in the quicklisp-client package. The resulting fasl would be
cached by asdf, and would cause errors if the system was later loaded in an
image that did not also have quicklisp loaded.

This converts the macros to functions, and uses run-time checks both to
determine whether the quicklisp package is available and also to look up
symbols in that package. It uses bare symbols to refer to symbol names, on
the off chance that someone is using a non-upcasing reader, and tries to
emit warnings when expected quicklisp symbols are not found (otherwise the
error messages tend to be "the function NIL is undefined").